dynamic decorator
